add asyImp of bigVector and finit state machine Push

+++ src/main/scala/org/apache/spark/asyspark/core/messages/server/logic/UniqueID.scala
@@ -3,4 +3,4 @@ package org.apache.spark.asyspark.core.messages.server.logic
 /**
   * Created by wjf on 16-9-25.
   */
-private[asyspark] case class UniqueID(id: Long)
+private[asyspark] case class UniqueID(id: Int)
